LayoutTests/inspector/timeline/timeline-layout.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/timeline-test.js"></script>
<script>

function performActions()
{
    var element = document.createElement("div");
    element.innerHTML = "Test data";
    document.body.appendChild(element);
    // Force layout.
    var foo = element.offsetHeight;
}

function test()
{
    InspectorTest.performActionsAndPrint("performActions()", "Layout");
}

</script>
</head>

<body onload="runTest()">
<p>
Tests the Timeline API instrumentation of a Layout event
</p>

</body>
</html>
