safari-534.53-branch/LayoutTests/inspector/timeline/timeline-layout-expected.txt

Tests the Timeline API instrumentation of a Layout event

Test data
Layout Properties:
{
    startTime : <number>
    data : {
    }
    children : <object>
    endTime : <number>
    type : "Layout"
    usedHeapSize : <number>
    totalHeapSize : <number>
}
